Map Update
This commit focuses on building the classes in mapupdate.h which at this
iteration are still very broken. Reorganizes a few of the functions in
other files, but nothing too important to comment on, because no other
files have been properly coded yet.
Adds a thorough readme covering all of the high points of the program
and each file as I believe they will function in the conclusion of this
project.